requirements.txt
fastapi
uvicorn
numpy
opencv-python-headless
pytesseract
pillow
python-dotenv